requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
channels-redis==4.2.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
daphne==4.1.2
django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
msgpack==1.1.1
packaging==25.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyopenssl==25.1.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==6.1.1
requests==2.32.4
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
twisted==25.5.0
txaio==23.1.1
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
websockets==13.1
zope.interface==7.2
